Title: 부분 문자열 중 팰린드롬의 수, Time: 136ms, Memory: 12MB, Status: Passed - CodeTree

+++ 241104/부분 문자열 중 팰린드롬의 수/counting-palindrome-substrings.java
@@ -39,7 +39,7 @@ public class Main {
 			}
 		}
 
-		int answer = 0;
+		long answer = 0;
 		for (int i = 0; i < n; i++) {
 			if (i % 2 == 0) {
 				answer += maxHalfLen[i] / 2;
